backend/requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-celery-beat==2.7.0
psycopg2-binary==2.9.10
redis==5.2.0
celery==5.4.0
python-decouple==3.8
Pillow==11.0.0
cryptography==44.0.3
requests==2.33.1
